Fixed NoGain SkillTries When SkillLvl is exactly 10 (#3314)

+++ src/vocation.cpp
@@ -182,10 +182,10 @@ uint64_t Vocation::getReqSkillTries(uint8_t skill, uint16_t level)
 	if (skill > SKILL_LAST) {
 		return 0;
 	}
-	return skillBase[skill] * std::pow(skillMultipliers[skill], level - 11);
+	return skillBase[skill] * std::pow(skillMultipliers[skill], static_cast<int32_t>(level - 11));
 }
 
 uint64_t Vocation::getReqMana(uint32_t magLevel)
 {
-	return 1600 * std::pow(manaMultiplier, magLevel - 1);
+	return 1600 * std::pow(manaMultiplier, static_cast<int32_t>(magLevel - 1));
 }
